Remove only frame_dummy from .init_array.
We handle init_array in two phases. In the first, we load the bytes of
the .init_array section in the data phase. In the second, we apply
relocations to symbolize the section so that it will re-link.

The symbol frame_dummy is included as part of glibc and so will be
included again when the program is re-compiled. This being the case,
retrowrite should not symbolize it. Since it appears in the .init_array
section it also needs to be removed there.

Previously, our logic was to ignore the first pointer of the init_array
section. In this commit, we modify that logic to explicitly look for a
function called frame_dummy regardless of what order it appears,
although it should appear first as it is included in crtBegin. This way
we are sure we are removing the correct function.

Signed-off-by: Antony Vennard <antony@vennard.ch>